Flip the container image to a UBI 9 image
Older Fedora releases move package registries to an archive after
having been EOL'd for some time which means the URLs change [1], [2].
Let's replace the old Fedora image with a UBI-9 where we should be
access the packages using the same URL for a long time.
Since the repo didn't list an rpms.in.yaml file, let's add one for
easier lockfile regenerating in the future.

[1] https://dl.fedoraproject.org/pub/fedora/linux/releases/39/
[2] https://dl.fedoraproject.org/pub/archive/fedora/linux/releases/39/

Signed-off-by: Erik Skultety <eskultet@redhat.com>