Add rpms.in.yaml
This is just a helper manifest file, because the resulting
rpms.lock.yaml based on the packages listed here would be actually
larger than what the next patch going to propose. The reason for that
being that this is for a scoped integration test, not an E2E, and so
the packages listed in this file were picked by hand based on tiny size
and they also happen to by system packages so the following CLI has
to be used to generate rpms.lock.yaml:

    $ rpm-lockfile-prototype \
        --bare \
        --outfile rpms.lock.yaml \
        rpms.in.yaml

Note this manifest IN file is based on CentOS Stream due to a
different package conservancy policy [1] (5 package builds vs only
latest).

[1] https://pkgs.org/

Signed-off-by: Erik Skultety <eskultet@redhat.com>